+++ cam.h
@@ -3,8 +3,8 @@
 struct fvec;
 
 void	 cam_move(int, float);
-void	 cam_revolve(struct fvec *, float, float);
-void	 cam_revolvefocus(float, float);
+void	 cam_revolve(struct fvec *, double , double);
+void	 cam_revolvefocus(double, double);
 void	 cam_rotate(const struct fvec *, const struct fvec *, int, int);
 void	 cam_roll(double);
 void	 cam_look(void);
